Create Column_Data_Types
This script identifies the data types of csv files in a directory, and creates and excel file with all the column headers and the data types. You can use this information in your codebooks or to determine whether you need to convert the data types. Make sure you change your file paths!